fix a bug, that crashed startup on new databases where old v1 admin name does not exist

+++ src/main/java/de/hsbo/kommonitor/datamanagement/config/InitialAccessControlSetup.java
@@ -156,7 +156,9 @@ public class InitialAccessControlSetup implements ApplicationListener<ContextRef
 		
 		// remove old administrator role explicitly
 		RolesEntity oldAdminRole = roleRepository.findByRoleName(oldAdministratorRoleName);
-		roleRepository.delete(oldAdminRole);
+		if(oldAdminRole != null) {
+			roleRepository.delete(oldAdminRole);
+		}
 		
 	}
